main.py
"""
PQC Migration Intelligence Agent Platform — Entry Point

Orchestrates the execution of specialized scanning agents and
coordinates with the Master Orchestrator for final report generation.
"""

import os
import sys
import logging

from dotenv import load_dotenv

load_dotenv()

# ─── Logging Setup ───────────────────────────────────────────────────────────
logging.basicConfig(
    level=os.getenv("LOG_LEVEL", "INFO"),
    format="%(asctime)s │ %(name)-20s │ %(levelname)-8s │ %(message)s",
    datefmt="%Y-%m-%d %H:%M:%S",
)
logger = logging.getLogger("pqc-migration")


def main():
    """Main entry point for the PQC Migration Agent Platform."""

    logger.info("=" * 60)
    logger.info("  PQC Migration Intelligence Agent Platform")
    logger.info("  Scanning for quantum-vulnerable cryptography...")
    logger.info("=" * 60)

    # ── 1. Initialize Agents ────────────────────────────────────────────
    from agents.web_api_agent import WebApiAgent
    from agents.iot_edge_agent import IoTEdgeAgent

    web_agent = WebApiAgent()
    iot_agent = IoTEdgeAgent()

    logger.info(f"Web/API Agent initialized  │ run_id={web_agent.run_id}")
    logger.info(f"IoT/Edge Agent initialized │ run_id={iot_agent.run_id}")

    # ── 2. Define Scan Targets ──────────────────────────────────────────
    # TODO: Load targets from configuration or command-line arguments
    # For now, these are placeholder targets for the demo
    web_targets = {
        "jwt_tokens": [],       # Add JWT tokens to scan
        "endpoints": [],        # Add API endpoints to scan
        "certificates": [],     # Add cert paths to scan
    }

    iot_targets = {
        "firmware_metadata": {},    # Add firmware metadata
        "device_profiles": [],      # Add device profiles
        "ota_config": {},           # Add OTA configuration
    }

    # ── 3. Run Scans ───────────────────────────────────────────────────
    logger.info("Starting scans...")

    # TODO: Uncomment when agent implementations are ready
    # web_assessment = web_agent.scan(web_targets)
    # web_agent.save_local()
    # logger.info(f"Web/API scan complete — {len(web_agent.findings)} findings")

    # iot_assessment = iot_agent.scan(iot_targets)
    # iot_agent.save_local()
    # logger.info(f"IoT/Edge scan complete — {len(iot_agent.findings)} findings")

    # ── 4. Orchestrate Final Report ─────────────────────────────────────
    # TODO: Invoke orchestrator (teammates' module)
    # from core.orchestrator import Orchestrator
    # orchestrator = Orchestrator()
    # orchestrator.collect_assessments()
    # report = orchestrator.generate_report()

    logger.info("=" * 60)
    logger.info("  Platform ready — implement agent scan() methods to begin")
    logger.info("=" * 60)


if __name__ == "__main__":
    main()
