requirements.txt
# AWS Strands Agent Framework
strands-agents
strands-agents-tools

# AWS SDK
boto3>=1.34.0

# Cryptography
cryptography>=42.0.0

# JWT / Token Analysis
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0

# Data Processing
pandas>=2.0.0

# Environment Management
python-dotenv>=1.0.0
